personal-expense-manager/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2
asyncpg
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
redis
httpx
pytest
pytest-asyncio
python-multipart
email-validator
aiosqlite
